Refactor all shared post components
