+++ web/src/app/api/gist.interface.d.ts
