+++ web/src/app/shared/post/content/content.component.html
@@ -1,0 +1,1 @@
+<article #content class="content"></article>
