+++ web/src/app/shared/post/content/index.ts
@@ -1,0 +1,1 @@
+export * from './content.component';
