+++ web/src/app/shared/post/excerpt/excerpt.component.html
@@ -1,0 +1,1 @@
+<article class="content" [innerHTML]="content"></article>
