+++ web/src/app/shared/post/excerpt/index.ts
@@ -1,0 +1,1 @@
+export * from './excerpt.component';
