--- web/src/app/shared/post/post-content.component.scss
@@ -1,112 +1,0 @@
-@import "../../../styles.scss";
-
-:host ::ng-deep {
-  color: #333;
-  font: normal 300 1.6rem Pridi, sans-serif;
-
-  a {
-    color: #0091ea;
-  }
-
-  blockquote {
-    border-left: 0.8rem solid #ececec;
-    display: block;
-    margin: 3.2rem 0;
-    padding: 0 1.6rem;
-
-    > p:first-child {
-      padding-top: 1.6rem;
-    }
-
-    > p:last-child {
-      padding-bottom: 1.6rem;
-    }
-  }
-
-  code,
-  pre {
-    font-family: "Roboto Mono", monospace;
-  }
-
-  h2,
-  h3,
-  h4 {
-    font-family: Athiti, sans-serif;
-    font-weight: 500;
-  }
-
-  h2 {
-    font-size: 3.6rem;
-  }
-  h3 {
-    font-size: 2.8rem;
-  }
-  h4 {
-    font-size: 1.8rem;
-  }
-
-  hr {
-    border: none;
-    border-bottom: 0.1rem solid #ececec;
-    display: block;
-    margin: 2.4rem auto;
-    padding-top: 0.8rem;
-    width: 100%;
-
-    @include md {
-      width: 51.4rem;
-    }
-  }
-
-  iframe {
-    border: none;
-    margin: 0.8rem 0;
-    max-width: 100%;
-    min-height: 32rem;
-    width: 100%;
-  }
-
-  img {
-    border-radius: 0.2rem;
-    display: block;
-    max-width: 100%;
-    margin-top: 3.2rem;
-  }
-
-  img + div.caption {
-    margin: 0.8rem 0 3.2rem 0;
-  }
-
-  p {
-    margin: 2.4rem 0;
-
-    &.fill-width {
-      margin: 0 -3.2rem !important;
-
-      @include sm {
-        margin: 0 calc((100vw / 2 * -1) + 25.6rem + 6.4rem) !important;
-        /*                                Sidebar   Padding */
-      }
-
-      @include lg {
-        margin: 0 calc((100vw / 2 * -1) + 25.6rem + 12.8rem) !important;
-        /*                                Sidebar   Padding  */
-      }
-
-      & > img {
-        border-radius: 0;
-        width: 100%;
-      }
-    }
-  }
-
-  pre {
-    margin: 3.2rem 0;
-  }
-
-  .caption {
-    color: #333;
-    font-size: 1.5rem;
-    text-align: center;
-  }
-}
