+++ web/src/app/shared/post/tags/index.ts
@@ -1,0 +1,1 @@
+export * from './tags.component';
