+++ web/src/app/shared/post/tags/tags.component.html
@@ -1,0 +1,5 @@
+<ul class="_list-unstyled">
+  <li *ngFor="let t of post.tags" class="tag-item">
+    <a [routerLink]="['/', 'tag', t.slug]">#{{t.name}}</a>
+  </li>
+</ul>
