+++ web/src/app/shared/post/title/title.component.html
@@ -1,0 +1,3 @@
+<a class="title" [routerLink]="isDisabledLink() ? null : href">
+  <ng-content></ng-content>
+</a>
