+++ web/src/app/shared/post/title/title.component.scss
@@ -1,0 +1,16 @@
+:host {
+  color: #333;
+  display: block;
+  font-family: Athiti, sans-serif;
+  font-weight: 500;
+}
+
+.title ::ng-deep h1 {
+  font-size: 4.8rem;
+  font-weight: inherit;
+}
+
+.title ::ng-deep h2 {
+  font-size: 3.6rem;
+  font-weight: inherit;
+}
